requirements.txt
aiocron==1.8
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
croniter==2.0.1
distlib==0.3.6
filelock==3.10.7
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
platformdirs==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
schedule==1.2.1
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.14
virtualenv==20.21.0
yarl==1.8.2
